+++ src/frontend/components/whatsNewDialog.tsx
@@ -9,7 +9,7 @@ interface IWhatsNewDialogProps {
 const WHATS_NEW_HEADER_TEXT = "Highlights from releases v1.92.57 and v1.92.56 include:";
 
 const WHATS_NEW_ITEMS = [
-  "Improved hidden feedback behavior to display blurred feedback while reading \"feedback blurred\".",
+  "Improved hidden feedback behavior to display blurred feedback while announcing \"feedback blurred\".",
   "Added user settings to toggle between show all teams or my teams and to toggle between scrolling by board or by column.",
   "Added team admin setting for configuring available work item types in \"Add work item\".",
   "Defaulted add-work-item type options to Requirement Backlog types for the Act tab and Focus mode.",