+++ src/frontend/components/feedbackCarousel.tsx
@@ -24,7 +24,7 @@ export interface FocusModeModel {
   nonHiddenWorkItemTypes: WorkItemType[];
   allWorkItemTypes: WorkItemType[];
   hideFeedbackItems: boolean;
-  isBoardOwner?: boolean;
+  canManageBoard?: boolean;
   activeTimerFeedbackItemId: string | null;
   onVoteCasted: () => void;
   requestTimerStart: (feedbackItemId: string) => void;
@@ -206,7 +206,7 @@ export const FeedbackCarousel: React.FC<IFeedbackCarouselProps> = ({ focusModeMo
           shouldHaveFocus: columnItem.shouldHaveFocus,
           hideFeedbackItems: focusModeModel.hideFeedbackItems,
           userIdRef: columnItem.feedbackItem.userIdRef,
-          isBoardOwner: focusModeModel.isBoardOwner,
+          canManageBoard: focusModeModel.canManageBoard,
           onVoteCasted: focusModeModel.onVoteCasted,
           groupCount: columnItem.feedbackItem.childFeedbackItemIds ? columnItem.feedbackItem.childFeedbackItemIds.length : 0,
           groupIds: columnItem.feedbackItem.childFeedbackItemIds ?? [],