+++ src/frontend/components/feedbackColumn.tsx
@@ -33,7 +33,7 @@ export interface FeedbackColumnProps {
   nonHiddenWorkItemTypes: WorkItemType[];
   allWorkItemTypes: WorkItemType[];
   isBoardAnonymous: boolean;
-  isBoardOwner?: boolean;
+  canManageBoard?: boolean;
   shouldFocusOnCreateFeedback: boolean;
   hideFeedbackItems: boolean;
   groupIds: string[];
@@ -120,7 +120,7 @@ export const createFeedbackItemProps = (columnProps: FeedbackColumnProps, column
     shouldHaveFocus: columnItem.shouldHaveFocus,
     hideFeedbackItems: columnProps.hideFeedbackItems,
     userIdRef: columnItem.feedbackItem.userIdRef,
-    isBoardOwner: columnProps.isBoardOwner,
+    canManageBoard: columnProps.canManageBoard ?? false,
     onVoteCasted: columnProps.onVoteCasted,
     groupCount: columnItem.feedbackItem.childFeedbackItemIds ? columnItem.feedbackItem.childFeedbackItemIds.length : 0,
     groupIds: columnItem.feedbackItem.childFeedbackItemIds ?? [],